fix: correcion carpetas controllers
