+++ src/routes/proyectosRoute.ts
@@ -1,0 +1,7 @@
+import { Router, Request, Response } from 'express'
+
+const router = Router()
+
+router.post('/', (_req: Request, _res: Response): Promise => {})
+
+export { router }
