+++ src/routes/userRoute.ts
@@ -1,9 +1,9 @@
+// import { Request, Response, Router } from 'express'
 import { Router } from 'express'
-import { loginUser } from '../controllers/User'
+import { loginUser } from '../controllers/user/User'
+
 const router = Router()
 
 router.post('/login', loginUser)
-
-//router.post('/register', registerUser)
 
 export { router }
